pyproject.toml
[project]
name = "ctrld-sync"
version = "0.1.0"
description = "Simple script which syncs defined Control D folders"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]
